pyproject.toml
[project]
name = "thinllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ujjwal Kumar", email = "kujjwal02@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.12.5",
    "jiter>=0.8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "coverage>=7.13.0",
    "ipykernel>=7.1.0",
    "langfuse>=3.11.2",
    "pyrefly>=0.46.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-html>=4.1.1",
    "pytest-mock>=3.15.1",
    "pytest-sugar>=1.1.1",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.9",
]
examples = [
    "streamlit>=1.52.2",
]
openai = [
    "openai>=2.12.0",
]
anthropic = [
    "anthropic>=0.42.0",
]
gemini = [
    "google-genai>=1.57.0",
]


[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.pyrefly]
python-version = "3.12.0"
ignore-missing-imports = ["*"]
project-includes = ["src/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: marks tests as unit tests (fast, mocked)",
    "integration: marks tests as integration tests (slower, real API calls)",
    "openai: marks tests specific to OpenAI provider",
    "anthropic: marks tests specific to Anthropic provider",
    "gemini: marks tests specific to Gemini provider",
]
# Show extra test summary info for failed tests
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--durations=0",
    "--cov=.",
    "--cov-report=xml:.test/coverage.xml",
    "--html=.test/report.html",
    "--self-contained-html",
]
